Refactor DeterminismTest to expose internal an application.
